requirements.txt
numpy
pandas
scikit-learn
joblib
nltk
TurkishStemmer
customtkinter
google-generativeai
python-dotenv